pyproject.toml
[project]
name = "ysjquiz"
version = "0.1.0"
description = ""
authors = [
    {name = "유승진",email = "91680691+sjgopromer2025@users.noreply.github.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi (>=0.115.12,<0.116.0)",
    "jinja2 (>=3.1.6,<4.0.0)",
    "uvicorn (>=0.34.0,<0.35.0)",
    "python-multipart (>=0.0.20,<0.0.21)",
    "psycopg2 (>=2.9.10,<3.0.0)",
    "sqlalchemy (>=2.0.40,<3.0.0)",
    "databases (>=0.9.0,<0.10.0)",
    "pydantic (>=2.11.1,<3.0.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "alembic (>=1.15.2,<2.0.0)",
    "bcrypt (>=4.3.0,<5.0.0)",
    "pyjwt (>=2.10.1,<3.0.0)"
]

[tool.poetry]
packages = [{include = "app"}]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
dev = "app.dev:main"


